// src/app/shared/model/ref-zipcode-obj.model.ts
// Generated by https://quicktype.io

export class RefZipcodeObj {
  RefZipcodeId: number;
  AreaCode2: string;
  AreaCode1: string;
  City: string;
  ZipcodeNumber: string;
  Zipcode: string;
  PhnArea: string;
  IsActive: boolean;
  SubZipcode: string;
  RefProvDistrictId: number;
  DistrictName: string;
  RowVersion: string;
  constructor() { this.RefZipcodeId = 0 }
}