// src/app/shared/model/ref-prov-district-obj.model.ts
export class RefProvDistrictObj {
    RefProvDistrictId: number;
    ProvDistrictCode: string;
    ProvDistrictName: string;
    DistrictRegRptCode: string;
    PhnArea : string;
    ParentId: number;
    Type: string;
    IsActive: boolean;
    RowVersion: string;
    constructor() { this.RefProvDistrictId = 0, this.RowVersion = "" }
}